requirements.txt
Django==4.0.5
psycopg2==2.9.3
python-dotenv==0.20.0
djangorestframework==3.13.1
coverage==6.4.1
django-filter==22.1
